show total usage in selected range
